Cleaner webpack configs
