+++ config/webpack.prod.js
@@ -1,0 +1,30 @@
+const webpack = require('webpack');
+const merge = require('webpack-merge');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
+const helpers = require('./helpers');
+const commonConfig = require('./webpack.common');
+
+module.exports = merge(commonConfig, {
+  output: {
+    filename: 'js/[name].[hash].js',
+    chunkFilename: '[id].[hash].chunk.js'
+  },
+
+  plugins: [
+    new webpack.optimize.DedupePlugin(),
+
+    new webpack.optimize.UglifyJsPlugin({
+      compressor: {
+        warnings: false,
+        screw_ie8: true
+      },
+      output: {
+        comments: false
+      }
+    }),
+
+    new CopyWebpackPlugin([{
+      from: helpers.root('client/public')
+    }])
+  ]
+});
