+++ webpack.config.js
@@ -1,0 +1,11 @@
+switch (process.env.NODE_ENV) {
+  case 'prod':
+  case 'production':
+    module.exports = require('./config/webpack.prod');
+    break;
+
+  case 'dev':
+  case 'development':
+  default:
+    module.exports = require('./config/webpack.dev');
+}
